requirements.txt
streamlit
numpy
pandas
altair
seaborn
matplotlib
tensorflow
Pillow